refactor: move execution_result.rs to crates/vm
