+++ crates/vm/backends/revm/mod.rs
@@ -4,12 +4,12 @@ pub mod helpers;
 #[cfg(feature = "l2")]
 mod mods;
 
-use super::execution_result::ExecutionResult;
 use super::BlockExecutionResult;
 use crate::constants::{
     BEACON_ROOTS_ADDRESS, CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS, HISTORY_STORAGE_ADDRESS,
     SYSTEM_ADDRESS, WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
 };
+use crate::execution_result::ExecutionResult;
 use crate::spec_id;
 use crate::EvmError;
 use db::EvmState;
